different home pages based on if not logged in or role. Rec -> All requests, regardless of status. Fab -> Only Active requests. Logged out -> Only completed requests, 'Done' status.

+++ src/controllers/comments/deletion.js
@@ -1,8 +1,13 @@
 const deleteComment = async (req, res) => {
   const {
+    session,
     db: { Comment },
     params: { id },
   } = req;
+
+  if (!session.userId) {
+    return res.sendStatus(401);
+  }
 
   const deleteComment = await Comment.deleteComment(id);
   if (deleteComment !== 1) return res.sendStatus(404);
